Lift the actions state up to Dashboard
